+++ CHANGELOG.md
@@ -1,13 +1,24 @@
 # Change log
 
-## v0.1.2 - 2024-04-08
+## v0.1.3 - 2024-06-28
+### Added
+#### UM Futures
+- `GET /fapi/v1/rateLimit/order`
+- `POST /fapi/v1/feeBurn`
+- `GET /fapi/v1/feeBurn`
+- `GET /fapi/v2/ticker/price`
+- `GET /futures/data/delivery-price`
 
 ### Changed
-- Update dependencies.
-- Resolve issue #3
+- Updated dependencies.
+
+## v0.1.2 - 2024-04-08
+### Changed
+- Updated dependencies.
+- Resolved issue #3
 
 ## v0.1.1 - 2023-10-20
-- Update dependencies.
+- Updated dependencies.
 
 ## v0.1.0 - 2023-10-13
 - First release.